Update to float

+++ src/MergeSort.java
@@ -4,12 +4,12 @@ import java.util.Arrays;
  * Created by chris on 2016-02-17.
  */
 public class MergeSort {
-    public static int[] sort(int[] a){
+    public static float[] sort(float[] a){
 
         if (a.length == 1) return a;
 
-        int[] b = new int[a.length/2];
-        int[] c = new int[a.length-b.length];
+        float[] b = new float[a.length/2];
+        float[] c = new float[a.length-b.length];
 
         System.arraycopy(a, 0, b, 0, b.length);
         System.arraycopy(a, b.length, c, 0, c.length);
@@ -20,7 +20,7 @@ public class MergeSort {
         return merge(b,c,a);
     }
 
-    private static int[] merge(int[] a, int[] b, int[]c){
+    private static float[] merge(float[] a, float[] b, float[]c){
         int indexa=0, indexb=0, indexc=0;
 
         while (indexa < a.length && indexb < b.length){
