+++ src/QuickSort.java
@@ -2,12 +2,12 @@
  * Created by chris on 2016-02-17.
  */
 public class QuickSort {
-    public static int[] sort(int[] a){
+    public static float[] sort(float[] a){
         quickSort(a, 0, a.length-1);
         return a;
     }
 
-    private static int[] quickSort(int[] a, int first, int last) {
+    private static float[] quickSort(float[] a, int first, int last) {
         if (first < last){
             int pivIndex = partition(a, first, last);
             quickSort(a, first, pivIndex-1);
@@ -16,8 +16,8 @@ public class QuickSort {
         return a;
     }
 
-    private static int partition(int[] a, int first, int last) {
-        int pivot = a[first];
+    private static int partition(float[] a, int first, int last) {
+        float pivot = a[first];
         int up = first;
         int down = last;
 
@@ -27,13 +27,13 @@ public class QuickSort {
             while (pivot < a[down] ) down--;
 
             if (up < down){
-                int tmp = a[up];
+                float tmp = a[up];
                 a[up] = a[down];
                 a[down] = tmp;
             }
         }while (up < down);
 
-        int tmp = a[first];
+        float tmp = a[first];
         a[first] = a[down];
         a[down] = tmp;
